+++ test/user_model.dart
@@ -1,0 +1,9 @@
+class UserModel {
+  String name;
+  int age;
+
+  UserModel(this.name, this.age);
+
+  @override
+  String toString() => 'UserModel(name: $name, age: $age)';
+}